@@ -1,7 +1,7 @@
 {
 	"name": "leedpdf",
 	"private": true,
-	"version": "2.22.0",
+	"version": "2.23.0",
 	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
@@ -25,31 +25,30 @@
 		"tauri:check": "cd src-tauri && cargo check"
 	},
 	"devDependencies": {
-		"@eslint/js": "^9.39.1",
+		"@eslint/js": "^9.39.2",
 		"@playwright/test": "^1.57.0",
-		"@rollup/rollup-darwin-arm64": "^4.53.3",
 		"@sveltejs/adapter-vercel": "^6.2.0",
 		"@sveltejs/enhanced-img": "^0.9.2",
-		"@sveltejs/kit": "^2.49.1",
+		"@sveltejs/kit": "^2.49.2",
 		"@sveltejs/vite-plugin-svelte": "^6.2.1",
 		"@tauri-apps/api": "^2.9.1",
-		"@tauri-apps/cli": "^2.9.5",
+		"@tauri-apps/cli": "^2.9.6",
 		"@testing-library/jest-dom": "^6.9.1",
 		"@types/node": "^20.12.7",
-		"@typescript-eslint/eslint-plugin": "^8.48.1",
-		"@typescript-eslint/parser": "^8.48.1",
+		"@typescript-eslint/eslint-plugin": "^8.49.0",
+		"@typescript-eslint/parser": "^8.49.0",
 		"@vitest/coverage-v8": "^3.2.4",
 		"@vitest/ui": "^3.2.4",
 		"eslint-plugin-svelte": "^3.13.1",
-		"jsdom": "^27.2.0",
+		"jsdom": "^27.3.0",
 		"prettier": "^3.7.4",
-		"prettier-plugin-svelte": "^3.4.0",
-		"svelte": "^5.45.6",
+		"prettier-plugin-svelte": "^3.4.1",
+		"svelte": "^5.46.0",
 		"svelte-check": "^4.3.4",
 		"svelte-eslint-parser": "^1.4.1",
 		"terser": "^5.44.1",
 		"typescript": "^5.9.3",
-		"vite": "7.2.7",
+		"vite": "7.3.0",
 		"vitest": "^3.2.4"
 	},
 	"pnpm": {
@@ -58,6 +57,7 @@
 			"cookie": "0.7.0"
 		},
 		"onlyBuiltDependencies": [
+			"core-js",
 			"esbuild",
 			"sharp"
 		]
@@ -73,15 +73,15 @@
 		"@types/markdown-it": "^14.1.2",
 		"@vercel/analytics": "^1.6.1",
 		"appwrite": "^21.5.0",
-		"autoprefixer": "^10.4.22",
+		"autoprefixer": "^10.4.23",
 		"jspdf": "^3.0.4",
 		"jszip": "^3.10.1",
-		"lucide-svelte": "^0.556.0",
+		"lucide-svelte": "^0.561.0",
 		"markdown-it": "^12.3.2",
 		"pdf-lib": "^1.17.1",
 		"pdfjs-dist": "5.4.449",
 		"postcss": "^8.5.6",
-		"posthog-js": "^1.302.2",
+		"posthog-js": "^1.306.1",
 		"tailwindcss": "^3.4.17"
 	}
 }
