@@ -44,6 +44,7 @@
 	let isFullscreen = false;
 	let showThumbnails = false;
 	let focusMode = false;
+	let presentationMode = false;
 	let isLoading = true;
 	let templateError = false;
 	let showShareModal = false;
@@ -614,6 +615,10 @@
 
 	function handleFullscreenChange() {
 		isFullscreen = !!document.fullscreenElement;
+		// Exit presentation mode when fullscreen is exited
+		if (!document.fullscreenElement && presentationMode) {
+			presentationMode = false;
+		}
 	}
 
 	function handleSharePDF() {
@@ -636,9 +641,18 @@
 		showShortcuts,
 		showThumbnails,
 		focusMode,
+		presentationMode,
 		onShowShortcutsChange: (value) => (showShortcuts = value),
 		onShowThumbnailsChange: (value) => (showThumbnails = value),
 		onFocusModeChange: (value) => (focusMode = value),
+		onPresentationModeChange: (value) => {
+			presentationMode = value;
+			if (value) {
+				document.documentElement.requestFullscreen?.();
+			} else {
+				document.exitFullscreen?.();
+			}
+		},
 		onFileUploadClick: handleFileUploadClick,
 		onStampToolClick: handleStampToolClick
 	}}
@@ -653,7 +667,7 @@
 	on:dragover={handleDragOver}
 	on:dragleave={handleDragLeave}
 >
-	{#if !focusMode}
+	{#if !focusMode && !presentationMode}
 		<Toolbar
 			onFileUpload={handleFileUpload}
 			onPreviousPage={() => pdfViewer?.previousPage()}
@@ -669,10 +683,19 @@
 			onSharePDF={handleSharePDF}
 			{showThumbnails}
 			onToggleThumbnails={handleToggleThumbnails}
+			{presentationMode}
+			onPresentationModeChange={(value) => {
+				presentationMode = value;
+				if (value) {
+					enterFullscreen();
+				} else if (document.fullscreenElement) {
+					exitFullscreen();
+				}
+			}}
 		/>
 	{/if}
 
-	<div class="w-full h-full" class:pt-12={!focusMode}>
+	<div class="w-full h-full" class:pt-12={!focusMode && !presentationMode}>
 		{#if templateError}
 			<!-- Error state when template is not found -->
 			<div class="h-full flex items-center justify-center">
@@ -708,12 +731,12 @@
 					<PageThumbnails isVisible={showThumbnails} onPageSelect={handlePageSelect} />
 				{/if}
 
-				<PDFViewer bind:this={pdfViewer} pdfFile={currentFile} />
+				<PDFViewer bind:this={pdfViewer} pdfFile={currentFile} {presentationMode} />
 			</div>
 		{/if}
 	</div>
 
-	{#if !focusMode}
+	{#if !focusMode && !presentationMode}
 		<HelpButton
 			position="absolute"
 			positionClasses="bottom-4 left-4"
@@ -724,7 +747,12 @@
 		<HomeButton {showThumbnails} />
 	{/if}
 
-	<Footer {focusMode} {getFormattedVersion} on:helpClick={() => (showShortcuts = true)} />
+	<Footer
+		{focusMode}
+		{presentationMode}
+		{getFormattedVersion}
+		on:helpClick={() => (showShortcuts = true)}
+	/>
 </main>
 
 <DragOverlay {dragOver} />
